# static/content/openapi.yaml
openapi: 3.1.0
info:
  title: Chronoverse HTTP API
  version: 1.0.0
  description: Public session-authenticated API exposed by the Chronoverse HTTP gateway.
servers:
  - url: http://localhost:8080
    description: Development
  - url: /api
    description: Production Nginx proxy
...
paths:
  /auth/register:
    post:
      operationId: registerUser
      summary: Register a user
      tags: [Authentication]
      description: Creates a user and issues session and CSRF cookies.
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AuthRequest'}
      responses:
        '201': {description: User registered and session created}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {$ref: '#/components/responses/Conflict'}
  /auth/login:
    post:
      operationId: loginUser
      summary: Log in
      tags: [Authentication]
      description: Validates credentials and issues session and CSRF cookies.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AuthRequest'}
      responses:
        '201': {description: Session created}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {description: Invalid email or password}
...
  /notifications:
    get:
      operationId: listNotifications
      summary: List notifications
      tags: [Notifications]
      security: [{cookieSession: []}]
      parameters: [{$ref: '#/components/parameters/Cursor'}]
      responses:
        '200':
          description: Paginated notifications
          content: {application/json: {schema: {$ref: '#/components/schemas/NotificationPage'}}}
    put:
      operationId: markNotificationsRead
      summary: Mark notifications read
      tags: [Notifications]
      security: [{cookieSession: [], csrfToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties: {ids: {type: array, minItems: 1, uniqueItems: true, items: {type: string, format: uuid}}}
      responses:
        '204': {description: Notifications marked read}
        '404': {description: One or more notifications do not exist or do not belong to the authenticated user; no notifications are changed}
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: session
      description: Encrypted session cookie issued after registration or login.
    csrfToken:
      type: apiKey
      in: cookie
      name: csrf
      description: CSRF cookie issued with the session and required for state-changing requests.
  parameters:
...
    Cursor: {name: cursor, in: query, required: false, schema: {type: string}, description: Opaque pagination cursor}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Valid UTF-8 containing 1-255 bytes after trimming only outer ASCII spaces (U+0020), with no Unicode control characters. Reuse a key only for retries of the exact same logical request.
      schema:
        type: string
        minLength: 1
        x-maxBytes: 255
  responses:
    BadRequest: {description: Invalid input or missing required idempotency key}
    Conflict: {description: Idempotency key was already used with a different logical request}
...
  schemas:
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, format: password, minLength: 8, maxLength: 72}
...
    Notification:
      type: object
      properties:
        id: {type: string, format: uuid}
        kind: {type: string}
        payload: {type: string, description: Notification-kind-specific JSON encoded as a string}
        read_at: {type: string, format: date-time}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    NotificationPage:
      type: object
      properties:
        notifications: {type: array, items: {$ref: '#/components/schemas/Notification'}}
        cursor: {type: string}
...